requirements.txt
ib_insync
mcp
streamlit
pandas
google-generativeai
python-dotenv
ccxt
